requirements.txt
alembic
bcrypt
cffi
click
coverage
Flask
Flask-Bcrypt
Flask-Cors
Flask-Migrate
Flask-Script
Flask-SQLAlchemy
Flask-Testing
itsdangerous
Jinja2
Mako
MarkupSafe
psycopg2
pycparser
PyJWT
python-editor
six
SQLAlchemy
Werkzeug==0.16.1
